Moved instances to AssociativeFlatten
