+++ src/main/scala/zio/prelude/AssociativeFlatten.scala
@@ -2,10 +2,12 @@ package zio.prelude
 
 import scala.annotation.implicitNotFound
 import scala.concurrent.Future
-import scala.util.Try
+import scala.util.{ Success, Try }
 
 import zio._
+import zio.prelude.NonEmptyList.single
 import zio.prelude.coherent.AssociativeFlattenCovariantDeriveEqual
+import zio.prelude.fx.ZPure
 import zio.stream.ZStream
 import zio.test.TestResult
 import zio.test.laws._
@@ -52,26 +54,31 @@ object AssociativeFlatten extends LawfulF.Covariant[AssociativeFlattenCovariantD
     associativeFlatten
 
   /**
-   * The `AssociativeFlatten` instance for `Cause`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Cause`.
    */
-  implicit val CauseAssociativeFlatten: AssociativeFlatten[Cause] =
-    new AssociativeFlatten[Cause] {
+  implicit val CauseIdentityFlatten: IdentityFlatten[Cause] =
+    new IdentityFlatten[Cause] {
+      override def any: Cause[Any] = Cause.fail(())
+
       override def flatten[A](ffa: Cause[Cause[A]]): Cause[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Chunk`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Chunk`.
    */
-  implicit val ChunkAssociativeFlatten: AssociativeFlatten[Chunk] =
-    new AssociativeFlatten[Chunk] {
+  implicit val ChunkIdentityFlatten: IdentityFlatten[Chunk] =
+    new IdentityFlatten[Chunk] {
+      def any: Chunk[Any] = Chunk.unit
+
       def flatten[A](ffa: Chunk[Chunk[A]]): Chunk[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Either`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Either`.
    */
-  implicit def EitherAssociativeFlatten[E]: AssociativeFlatten[({ type lambda[+a] = Either[E, a] })#lambda] =
-    new AssociativeFlatten[({ type lambda[+a] = Either[E, a] })#lambda] {
+  implicit def EitherIdentityFlatten[E]: IdentityFlatten[({ type lambda[+a] = Either[E, a] })#lambda] =
+    new IdentityFlatten[({ type lambda[+a] = Either[E, a] })#lambda] {
+      def any: Either[E, Any] = Right(())
 
       def flatten[A](ffa: Either[E, Either[E, A]]): Either[E, A] =
         ffa match {
@@ -82,34 +89,42 @@ object AssociativeFlatten extends LawfulF.Covariant[AssociativeFlattenCovariantD
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Exit`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Exit`.
    */
-  implicit def ExitAssociativeFlatten[E]: AssociativeFlatten[({ type lambda[+a] = Exit[E, a] })#lambda] =
-    new AssociativeFlatten[({ type lambda[+a] = Exit[E, a] })#lambda] {
+  implicit def ExitIdentityFlatten[E]: IdentityFlatten[({ type lambda[+a] = Exit[E, a] })#lambda] =
+    new IdentityFlatten[({ type lambda[+a] = Exit[E, a] })#lambda] {
+      def any: Exit[E, Any] = Exit.unit
+
       def flatten[A](ffa: Exit[E, Exit[E, A]]): Exit[E, A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Future`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Future`.
    */
-  implicit val FutureAssociativeFlatten: AssociativeFlatten[Future] =
-    new AssociativeFlatten[Future] {
+  implicit val FutureIdentityFlatten: IdentityFlatten[Future] =
+    new IdentityFlatten[Future] {
+      def any: Future[Any] = Future.successful(())
+
       def flatten[A](ffa: Future[Future[A]]): Future[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Id`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Id`.
    */
-  implicit val IdAssociativeFlatten: AssociativeFlatten[Id] =
-    new AssociativeFlatten[Id] {
+  implicit val IdIdentityFlatten: IdentityFlatten[Id] =
+    new IdentityFlatten[Id] {
+      def any: Id[Any] = Id(())
+
       def flatten[A](ffa: Id[Id[A]]): Id[A] = Id.unwrap(ffa)
     }
 
   /**
-   * The `AssociativeFlatten` instance for `List`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `List`.
    */
-  implicit val ListAssociativeFlatten: AssociativeFlatten[List] =
-    new AssociativeFlatten[List] {
+  implicit val ListIdentityFlatten: IdentityFlatten[List] =
+    new IdentityFlatten[List] {
+      def any: List[Any] = List(())
+
       def flatten[A](ffa: List[List[A]]): List[A] = ffa.flatten
     }
 
@@ -123,59 +138,93 @@ object AssociativeFlatten extends LawfulF.Covariant[AssociativeFlattenCovariantD
     }
 
   /**
-   * The `AssociativeFlatten` instance for `NonEmptyChunk`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `NonEmptyChunk`.
    */
-  implicit val NonEmptyChunkAssociativeFlatten: AssociativeFlatten[NonEmptyChunk] =
-    new AssociativeFlatten[NonEmptyChunk] {
+  implicit val NonEmptyChunkIdentityFlatten: IdentityFlatten[NonEmptyChunk] =
+    new IdentityFlatten[NonEmptyChunk] {
+      def any: NonEmptyChunk[Any] = NonEmptyChunk.single(())
+
       def flatten[A](ffa: NonEmptyChunk[NonEmptyChunk[A]]): NonEmptyChunk[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Option`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Option`.
    */
-  implicit val OptionAssociativeFlatten: AssociativeFlatten[Option] =
-    new AssociativeFlatten[Option] {
+  implicit val OptionIdentityFlatten: IdentityFlatten[Option] =
+    new IdentityFlatten[Option] {
+      def any: Option[Any] = Some(())
+
       def flatten[A](ffa: Option[Option[A]]): Option[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Try`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Try`.
    */
-  implicit val TryAssociativeFlatten: AssociativeFlatten[Try] =
-    new AssociativeFlatten[Try] {
+  implicit val TryIdentityFlatten: IdentityFlatten[Try] =
+    new IdentityFlatten[Try] {
+      def any: Try[Any] = Success(())
+
       def flatten[A](ffa: Try[Try[A]]): Try[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `Vector`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `Vector`.
    */
-  implicit val VectorAssociativeFlatten: AssociativeFlatten[Vector] =
-    new AssociativeFlatten[Vector] {
+  implicit val VectorIdentityFlatten: IdentityFlatten[Vector] =
+    new IdentityFlatten[Vector] {
+      def any: Vector[Any] = Vector(())
+
       def flatten[A](ffa: Vector[Vector[A]]): Vector[A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `ZIO`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `ZIO`.
    */
-  implicit def ZIOAssociativeFlatten[R, E]: AssociativeFlatten[({ type lambda[+a] = ZIO[R, E, a] })#lambda] =
-    new AssociativeFlatten[({ type lambda[+a] = ZIO[R, E, a] })#lambda] {
+  implicit def ZIOIdentityFlatten[R, E]: IdentityFlatten[({ type lambda[+a] = ZIO[R, E, a] })#lambda] =
+    new IdentityFlatten[({ type lambda[+a] = ZIO[R, E, a] })#lambda] {
+      def any: ZIO[R, E, Any] = ZIO.unit
+
       def flatten[A](ffa: ZIO[R, E, ZIO[R, E, A]]): ZIO[R, E, A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `ZManaged`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `ZManaged`.
    */
-  implicit def ZManagedAssociativeFlatten[R, E]: AssociativeFlatten[({ type lambda[+a] = ZManaged[R, E, a] })#lambda] =
-    new AssociativeFlatten[({ type lambda[+a] = ZManaged[R, E, a] })#lambda] {
+  implicit def ZManagedIdentityFlatten[R, E]: IdentityFlatten[({ type lambda[+a] = ZManaged[R, E, a] })#lambda] =
+    new IdentityFlatten[({ type lambda[+a] = ZManaged[R, E, a] })#lambda] {
+      def any: ZManaged[R, E, Any] = ZManaged.unit
+
       def flatten[A](ffa: ZManaged[R, E, ZManaged[R, E, A]]): ZManaged[R, E, A] = ffa.flatten
     }
 
   /**
-   * The `AssociativeFlatten` instance for `ZStream`.
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `ZStream`.
    */
-  implicit def ZStreamAssociativeFlatten[R, E]: AssociativeFlatten[({ type lambda[+a] = ZStream[R, E, a] })#lambda] =
-    new AssociativeFlatten[({ type lambda[+a] = ZStream[R, E, a] })#lambda] {
+  implicit def ZStreamIdentityFlatten[R, E]: IdentityFlatten[({ type lambda[+a] = ZStream[R, E, a] })#lambda] =
+    new IdentityFlatten[({ type lambda[+a] = ZStream[R, E, a] })#lambda] {
+      def any: ZStream[R, E, Any] = ZStream.unit
+
       def flatten[A](ffa: ZStream[R, E, ZStream[R, E, A]]): ZStream[R, E, A] = ffa.flatten
+    }
+
+  /**
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `NonEmptyList`.
+   */
+  implicit val NonEmptyListIdentityFlatten: IdentityFlatten[NonEmptyList] =
+    new IdentityFlatten[NonEmptyList] {
+      val any: NonEmptyList[Any] = single(())
+
+      def flatten[A](ffa: NonEmptyList[NonEmptyList[A]]): NonEmptyList[A] = ffa.flatten
+    }
+
+  /**
+   * The `AssociativeFlatten` and `IdentityFlatten` instance for `ZPure`.
+   */
+  implicit def ZPureIdentityFlatten[S, R, E]: IdentityFlatten[({ type lambda[+A] = ZPure[S, S, R, E, A] })#lambda] =
+    new IdentityFlatten[({ type lambda[+A] = ZPure[S, S, R, E, A] })#lambda] {
+      def any: ZPure[S, S, Any, Nothing, Any] = ZPure.unit
+
+      def flatten[A](ffa: ZPure[S, S, R, E, ZPure[S, S, R, E, A]]): ZPure[S, S, R, E, A] = ffa.flatten
     }
 
 }
